+++ FirebaseInAppMessaging/Sources/DefaultUI/Banner/FIRIAMBannerViewController.m
@@ -142,6 +142,7 @@ static const CGFloat kSwipeUpThreshold = -10.0f;
       }
     }
     self.imageView.image = image;
+    self.imageView.accessibilityLabel = self.inAppMessage.campaignInfo.campaignName;
   } else {
     // Hide image and remove the bottom constraint between body label and image view.
     self.imageViewWidthConstraint.constant = 0;
@@ -286,6 +287,9 @@ static const CGFloat kSwipeUpThreshold = -10.0f;
                      self.view.center = normalCenterPoint;
                    }
                    completion:nil];
+
+  // Announce via VoiceOver that the banner has appeared. Highlight the title label.
+  UIAccessibilityPostNotification(UIAccessibilityScreenChangedNotification, self.titleLabel);
 }
 
 - (void)setupAutoDismissTimer {