+++ FirebasePerformance/Sources/FPRClient.m
@@ -25,7 +25,7 @@
 #import "FirebasePerformance/Sources/FPRConsoleLogger.h"
 #import "FirebasePerformance/Sources/FPRProtoUtils.h"
 #import "FirebasePerformance/Sources/Instrumentation/FPRInstrumentation.h"
-#import "FirebasePerformance/Sources/Loggers/FPRGDTCCLogger.h"
+#import "FirebasePerformance/Sources/Loggers/FPRGDTLogger.h"
 #import "FirebasePerformance/Sources/Timer/FIRTrace+Internal.h"
 #import "FirebasePerformance/Sources/Timer/FIRTrace+Private.h"
 
@@ -117,10 +117,12 @@
 
   dispatch_group_async(self.eventsQueueGroup, self.eventsQueue, ^{
     // Create the Logger for the Perf SDK events to be sent to Google Data Transport.
-    self.gdtLogger = [[FPRGDTCCLogger alloc] initWithLogSource:logSource];
+    self.gdtLogger = [[FPRGDTLogger alloc] initWithLogSource:logSource];
 
+#ifdef TARGET_HAS_MOBILE_CONNECTIVITY
     // Create telephony network information object ahead of time to avoid runtime delays.
     FPRNetworkInfo();
+#endif
 
     // Update the configuration flags.
     [self.configuration update];