+++ FirebaseDynamicLinks/CHANGELOG.md
@@ -1,5 +1,5 @@
 # v7.7.0
-- [added] Adding "utmParametersDictionary" api to FIRDynamicLink. (#6730)
+- [added] Added `utmParametersDictionary` property to `DynamicLink`. (#6730)
 
 # v7.6.0
 - [fixed] Fixed build warnings introduced with Xcode 12.5. (#7434)